New feature:	External C/C++ dependencies can be bundled up for easy 		redistribution.
In this first stage of implementing this feature for the UNIX build,
the dependencies are *not* installed by default when making an
installation or a release.  Instead, there is a separate target named
'install-deps' that must be run to install dependencies.  By default,
dependencies install into the same directory tree as VR Juggler, but
this can be changed by setting the variable deps_prefx on the command
line when running GNU make.

For the developer installation (the instlinks stuff), the dependecies
are installed into the instlinks tree.  Since the *-config scripts have
not yet been updated, the only real effect of this is that developers
can set $LD_LIBRARY_PATH to point only to $VJ_BASE_DIR/lib and get all
the shared libraries resolved.  The updates to the *-config scripts are
the next stage.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@14727 08b38cba-cd3b-11de-854e-f91c5b6e4272