Supporting both static and dynamic libraries on Windows is rather tricky. The ways I have seen it done require that user build processes require passing an extra /D option to Visual C++ that identifies that DLL linking or static linking will be used.  This seems very error prone to me.
Since we do not currently have the infrastructure to support that anyway,
it seems to me that we won't be supporting static libraries on Windows any
time soon.  As such, I have renamed all the configurations to remove the
DLL distinction.  Now, everything is either "Debug" or "Release."  Besides
all of the above, this change simplifies the use of the Visual Studio
solution on the user end.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@14954 08b38cba-cd3b-11de-854e-f91c5b6e4272