Changed the visibility of the pointer to the native C++ memory for a given object so that it can be acquired from a public read-only property. Prevoiusly, the visibility of the raw pointer value only allowed access from within the same assembly, but this proved to be too restrictive.  The intention is to get this code compiling with Microsoft's C# compiler.
This is a pretty significant change aimed at fixing a serious problem.  I
had originally hoped that custom marshalers would be the silver bullet in
terms of dealing with accessing the raw pointer value, but now I see that
that is not a viable solution.  The existance of custom marshalers allows
any code to get to the raw pointer value, so there might as well be a
publicly accessible property.  The value cannot be changed once is set,
however, so there is still some assurance of data integrity.  The problem
is that any code can now hang onto the raw memory value and try to use it.
Of course, that was possible before by using the method
MarshalManagedToNative() in any custom marshaler.  Nevertheless, the
long-term effects of this change could be serious...


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@15214 08b38cba-cd3b-11de-854e-f91c5b6e4272