Bug fixed:	If a visual was requested with stereo enabled but none 		could be found, the wrong attributes were being disabled. 		There was an off-by-one error causing this to occur.
This revision is a small effort to clean up the feature-disabling code
used as a fallback when a GLX visual cannot be found that meets the
user's request.  The basic issue is how to change the vector viattrib
in getGlxVisInfo() without requiring a lot of effort to keep track of
the vector's changing size and such.  The approach had been to record
indices of features that could be disabled, but there one case of
hard-coded badness based on the contents of the vector at the time the
code was written.  Now, all the indices are recorded as the vector is
built up to make the code a little more flexible and more robust in the
face of future changes.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@15321 08b38cba-cd3b-11de-854e-f91c5b6e4272