Initial version of the code needed to allow graphical resizing and placement of surface and simulator viewports within a display window.  This isn't 100% finished yet, but it is working.
The classes in the 'placer' package duplicate code that Ben wrote two years
ago.  Amazingly, I don't think I had to modify anythin in that directory.
If this placer code is going to be utilized by multiple editors, it
would probably be good to migrate it into a collection of common UI
classes.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@15374 08b38cba-cd3b-11de-854e-f91c5b6e4272