Disable the mouse cursor change when the user elects to hide the mouse pointer when inside the display window being represented by a DisplayWindowFrame instance.  I hope that this will be a temporary change, but right now, it is fighting with the placer's desire to change the mouse cursor for viewport move and resize events.
The placer now changes the cursor of the root pane's glass pane.  In order
for the context menus to work, the root pane's glass pane must be visible,
but that prevents child components from effecting a visual change on the
mouse cursor.  (They can still call setCursor(), but it doesn't actually
change anything in the UI.)  In order to allow the placer to change the
cursor for move and resize events, it has to change the glass pane's cursor
rather than its own.  This may actually be a bad idea, but it is what I
have found so far in searching Java developer discussions.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15394 08b38cba-cd3b-11de-854e-f91c5b6e4272