Fixed minor coding standard violations.  Wrapped a long line or two. Fixed Doxygen formatting errors and removed unneeded HTML tags in comments.
Updated comments to reflect reality.  In particular, I removed the part of
the getDataVector() documentation that says that the value returned can be
used in "pre-draw" (referring to preDraw(), which was renamed to preFrame()
in January 2000).  As far as I can tell, the value returned by
getDataVector() cannot be used anywhere but in a rendering thread.  That
makes it pretty useless (IMO) since that's where context-specific data
access is handled by the smart pointer semantics.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15514 08b38cba-cd3b-11de-854e-f91c5b6e4272