Refactor:	Collect the information related to the environment 		variables needed for Visual Studio to do its job into a 		simple class.
This is Step 1 in a larger effort to eliminate duplicated code introduced
as a result of having two paths through the code (the text-based interface
and the Tkinter GUI).  The primary goal here is to minimize the effort
required to manage the options related to environment variables.  An added
benefit of this change is that the processing of input in the text-based
interface and the setup of the text fields in the Tk GUI can be done using
loops insteda of manual enumeration.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15518 08b38cba-cd3b-11de-854e-f91c5b6e4272