Obtained from:	wzchang
collideApp is an app that Wilson Chang wrote over the summer to
show the Performance Monitoring at work.  It has a balls bouncing
aournd in a rectangle which you can add or subtract from.  The idea
being to watch the CPU time required to calculate the collisions go
up and down.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15625 08b38cba-cd3b-11de-854e-f91c5b6e4272