New feature:	Added the concept of "editable enumerations" to property 		definitions.  These direct VRJConfig to use editble 		JComboBox objects for properties with enumerated allowed 		values.  Because they are editable, the user can enter 		his or her own value when necessary.
The motivation for this new feature is to handle the case when a
property value can be one of a pre-defined set or a user-defined value.
This is most likely to come up in the case of plug-ins.

This change introduces a small modification to the .jdef file format.
The .jdef file format version has been incremented to 3.1, and the XML
Schema for that format has been updated to reflect the difference.  A
transform (jdef_3.0-3.1.xsl) has been written to automate the transition
for anyone with custom config element types.  See RELEASE_NOTES.txt for
more details.

The version has been incremented to 0.18.0 to mark this change.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15654 08b38cba-cd3b-11de-854e-f91c5b6e4272