Bug fixed:	If the client side of the communication with the MotionStar 		got out of sync with the message sequence number, it did 		not recover correctly.
The basic problem here was that the internal sequence number was not
incremented after resynchronizing with the server's sequence number.  Thus,
once it got out of sync, it would always be off by one.

NOTE: This is not a serious bug!  It just resulted in a lot of output if
the client's expected sequence number did not match the received sequence
number.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15699 08b38cba-cd3b-11de-854e-f91c5b6e4272