After working with Boost on Mac OS X some more, I discovered that a correct installation of Boost 1.31 (one that actually has shared libraries compiled and installed) does not include the toolset name ('darwin') in the file names of the compiled libraries.  That makes things a little more complicated.
This change addresses the issue by slightly changing the way that the
toolset name is used to compose the Boost.Filesystem library name.  The
most important aspect of this change is that the Boost version number is
now included as part of the library name against which VPR will be linked.
This is necessary because on Darwin (OS X) and on Windows, there are no
dynamically loadable Boost libraries whose names do not contain the Boost
version number.  With this change, we are now linking against a uniformly
named shared library on *all* platforms.  With the upcoming Boost 1.32
release, this will probably prove to be beneficial and necessary anyway.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15716 08b38cba-cd3b-11de-854e-f91c5b6e4272