New feature: SF Bug#: [971282] Cannot deselect a value in a JComboBox
You can now select "None" as a valid option while editing a ConfigElementPointer.
This is only a representation of an empty string behind the scenes. This is
possible because of the great wisdom of Ben Scott when he decided to use the
PropertyEditors that came with Java, it took longger to find this bug on SF than
it did to actually fix it.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15820 08b38cba-cd3b-11de-854e-f91c5b6e4272