New feature:	vpr::SempahorePosix allocates a named semaphore on 		platforms where unnamed semaphores are not supported.
At this time, only versions of Darwin up through 7.5 (OS X 10.3,
"Panther") make use of named semaphores.  It is possible that this
would address a problem I ran into in 1998 with HP-UX 10.20, but I
have no way of testing that theory.  It's too bad I didn't have
Google in 1998.

This mimics the behavior of NSPR where named semaphores are the only
type used.

Bumped version to 0.40.1.  This version now fully eliminates the need
for NSPR on Mac OS X.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15826 08b38cba-cd3b-11de-854e-f91c5b6e4272