SF Bug#: [1047388] Files from a multi-file config are not listed in context.
Moved the location of when the ContextToolbar's GUI was initialized with jbInit()
to be before setting the ConfigContext passed to the constructor. This bug was
actually caused because the expand button that brings up the EditContextPopup
was being enabled in this.setConfigContext(), but was being disabled in jbInit()
because it is the default. Reordering these calls fixes the bug.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15884 08b38cba-cd3b-11de-854e-f91c5b6e4272