Bug fixed: ConfigeElements where not being updated after undo/redo events. This is caused            because they were not receiving ConfigElementEvents when a property value changed.
In order to fix this I simply had to make sure that I was passing the old_value into
ConfigElement.firePropertyValeChanged() instead of the new value. This might cause problems
with the bug fix Patrick made. Either way when a property value is changed it will currently
be changed twice at least, once when it is actually changed and once on recieving the
ConfigElementEvent.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@15892 08b38cba-cd3b-11de-854e-f91c5b6e4272