Refactor:	Remove framecounter code from ProfileManager.
This counter has never really worked well in here and doesn't work the way a user would expect.  It also will completely break with multithreading.
So now it is gone.

You can get the same information by either keeping a clock yourself or adding a Profile sample that encompasses the entire frame.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@15942 08b38cba-cd3b-11de-854e-f91c5b6e4272