New feature:	Multi-threaded profile data collection and management.
This changes the profile manager to store all profile root nodes in thread specific data.
This gives each thread a separate profile tree to collect information within.

printTree has been expanded so that it will print the data just the current thread or all threads.

NOTE: The printing of all threads seg faults right now in the test case.  I believe this is because there is a bug in thre ThreadManager or
Thread system in general that allows invalid threads to be in the ThreadManager.  This needs tracked down further.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15944 08b38cba-cd3b-11de-854e-f91c5b6e4272