Profit from the genius of the Boost developers and contributors to get the Mac OS X shared library linking process correct--finally.  With this change, the PyJuggler extension modules load without getting stuck on the dreaded semaphore_wait_trap() call.
This is a brute-force approach to fixing shared library creation since it
is a two-step process to do it correctly (as far as I can tell).  No other
platform requies that, so I don't have a good way to generalize it so that
it can be encapsulated within the configure script.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@15965 08b38cba-cd3b-11de-854e-f91c5b6e4272