Bug fixed: Mouse Centering on locked keys SF Bug#:1056317
Extracted data from the graphics window for size and width if it is
not a pure event window since it will have bogus info on size and origin.
Also added ability to catch resize events for event windows
so they will find new center.

Tested GL and Performer.

Notes: On Performer all events seem to be caught except for mouse clicks
      this was a condition that existed before this addition.

      Also, You cannot hide the cursor if the event source is a graphics
      window because you do not have valid information about the display
      and window that are needed to create the empty cursor.  One way
      to fix this is to add the *Display to WindowInfo but you must wait
      until the window has been created and opened....This may not work
      for Performer though since I have not found a way to get the
      *Display handle for the X Window from Performer.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@16009 08b38cba-cd3b-11de-854e-f91c5b6e4272