New feature:	Allow loading of the remote run-time reconfiguration 		plug-in to be disabled by setting an environment variable 		(NO_RTRC_PLUGIN).
The primary reason I have added this is related to the use of PyJuggler
on Mac OS X.  When the Python interpreter is the application into which
symbols are being loaded, static initialization within omniORB is not
happening correctly, even with DYLD_BIND_AT_LAUNCH set (the technique
suggested by users of omniORBpy on Mac OS X).  I may still be doing
someting wrong, but I don't know what it might be at this time.

The reason I actually thought of adding this, however, is for the IRIX
SPROC case.  On IRIX, failure to load a shared library results in a pretty
ugly message that includes the phrase "fatal error."  People don't seem to
read error messages very closely, but they *will* pick out that phrase and
assume it has something to do with any little thing that might go wrong
with their application's exection.  By providing a means to disable
loading of a plug-in that cannot be used with SPROC threads anyway, we can
give users a modicum of peace of mind.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@16017 08b38cba-cd3b-11de-854e-f91c5b6e4272