SF Bug#: [809067] VRJConfig allows multiple same-named elements.
New ConfigElements are now named based on their definitions name
and a unique number that starts and 0 and increases until a unique
name is found. Here is an example for a new "Cluster Manager Configuration.

ClusterManagerConfiguration0
ClusterManagerConfiguration1
ClusterManagerConfiguration2

Also if you try to rename a existing ConfigElement that matches
another elements name a dialog box is displayed and the name is
set back to the original.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@16067 08b38cba-cd3b-11de-854e-f91c5b6e4272