New feature:	The EnvironmentService interface has a new method, 		getenv(), that allows Java code to read environment 		variable values from the run-time environment.  This is 		implemented as a native method that calls getenv(3).
Using JNI to implement this is not exactly my favorite idea, but since
System.getenv() went away (but is back for JDK 1.5), this is the only
way I can find to make this happen.  In the larger scheme of things, this
method is extremely valuable since the use of Java system properties to
pass in environment variables is getting out of hand.

Bumped version to 0.19.2 to note this API extension.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@16070 08b38cba-cd3b-11de-854e-f91c5b6e4272