SF Bug#: [ 969928 ] vrjconfig: select file broken in context w/ multiple files.          [ 971283 ] No pop-up message when trying to edit read-only data source.
First bug turned out to be two calls to itr.next() within a for loop. Simple fix  :)

All editor controls are now disabled when they are editing a ConfigElement that is
in a read only DataSource. Also the Add/Remove buttons in the ConfigContextEditor
now are only enabled if the current ConfigContext has at least one DataSource that
is writable. Also when there are multiple DataSources in a ConfigContext and a
ConfigElement is added, only writable DataSources are presented in the list of
DataSources to add the new ConfigElement to.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@16107 08b38cba-cd3b-11de-854e-f91c5b6e4272