Refactor:	The determiniation about when to make a new version of a 		config definition has been moved from the Config Broker 		to the config definition editor.  Previously, every time 		a config definition was saved by the Config Broker, a new 		version was created.  Now, the config definition editor 		knows when a new version is needed and takes steps to make 		that happen. Bug fixed:	The config definition editor simply didn't work.  It 		could not be used to create new definitions (a 		long-standing problem), and it could not handle version 		changes when property definitions were added or removed 		from an existing config definition. SF Bug#:	625667
The refactoring included with this commit should allow us to make a better
decision about whether upgrade transforms can be created automatically or
not.  Prior to this, the determination about when to create a new config
definition version was entirely in the hands of the Config Broker, but it
knew nothing about what had changed in the config definition.  Now that the
decision is made by the config definition editor, it may be possible to use
the knowledge of what has changed in the config definition to determine how
to create an upgrade transform (or at least the beginnings of one).


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@16160 08b38cba-cd3b-11de-854e-f91c5b6e4272