Removed the notion in the Java API that a property definition has a human-readable name.  The .jdef XML file format does not include any such attribute.  The name of a property definition is its token, which is the XML element name that appears in .jconf files.
This change includes a simplification to the property definition editor.
Now, there is no need to edit both the name and the token.  Changing the
(conceptual) name of a property definition results in its token being
updated.  The editor takes care of converting the (conceptual) name to a
valid XML element name.

Bumped version to 0.23.0.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@16163 08b38cba-cd3b-11de-854e-f91c5b6e4272