Initial version of a transform for handling the configuration changes that cannot be expressed by the config definition upgrade transforms.  This is based on the current state of code that Aron has not committed.  It also includes code for handling the renaming of the machine_specific element, which has not actually been done yet.  This code is commented out and will remain so until the actual renaming takes place.
NOTE: I have included some code for converting the "old" version 4 of the
display_window config element type to the "new" version 4.  This basically
works, but there is no way to handle the case of something like
standalone.jconf.  That sort of configuration will have to be updated by
hand.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@16272 08b38cba-cd3b-11de-854e-f91c5b6e4272