Renamed the config element type sound_manager to sound_manager_sonix. While this isn't the most critical change in the world, it occurred to me that the config element for the Sonix Sound Manager should have a name that reflects its actual use.
As part of this change, I extended the "api" property in the
sound_mangaer_sonix config definition to take advantage of editable
enumerations.  In this way, users configuring the Sonix Sound Manager can
choose the available sound APIs from a drop-down list instead of having to
know themselves the names of the Sonix plug-ins.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@16278 08b38cba-cd3b-11de-854e-f91c5b6e4272