New feature:	Config definition versions can now be marked as "abstract" 		to indicate to editor tools that that definition version 		cannot be instantiated as a config element.  This will 		help prevent invalid configurations from being created by 		users who try to create config elements of element base 		types.
While this change does introduce a new child of the XML element
<definition_version>, I have implemented this in a way that is backwards
compatible with versions 3.0 and 3.1 of the config definition file format.
Hence, there is no need to increment the file format version, but the XML
schema definition.xsd will have to be updated on the VR Juggler website.

This includes an API change to org.vrjuggler.jccl.config.ConfigDefinition
and to org.vrjuggler.jccl.config.event.ConfigDefinitionListener.  Bumped
version to 0.24.0 to mark this change.

Approved by:	aronb


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@16284 08b38cba-cd3b-11de-854e-f91c5b6e4272