Bug fixed:	The MotionStar backpack had to be turned on in order for 		the driver to be able to configure everything correctly. SF Bug#:	1067821
The problem here turned out to be quite simple.  The driver was telling
the MotionStar chassis to disable any birds that did not have sensors
attached.  When the pack is off, no birds are reported as having sensors,
and when it was turned on, the server did as it was told and disabled the
all the birds--even those with sensors.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@16375 08b38cba-cd3b-11de-854e-f91c5b6e4272