New feature:	Graphical display of live performance data from Kernel.
This is just a very simple/stupid demo of the ability that we now have to display live performance data coming from
inside VR Juggler.  This is more a proof-of-concept for everyone else to see then a final product.  I hope that we can
refine this code to make a better display in the future.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@16435 08b38cba-cd3b-11de-854e-f91c5b6e4272