Bug fixed:	POSIX CondVar time over run bug found and fixed
The old implementation of condvar did not take into account that adding the current time and a wait time could end up with more nano-seconds
in the timespec struct then there are in a second.  The new code accounts for this and rolls full seconds into the second part of the struct.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@16513 08b38cba-cd3b-11de-854e-f91c5b6e4272