A few major fixes:
   * Ignore the user-provided minimum and maximum analog values.  The
     standalone device driver sets that itself through Direct Input.
   * Making the above change results in proper scaling of the analog stick
     data.  Now, the analog data behaves as expected (and desired).
   * Removed hard-coded values for the number of axes and the number of
     buttons.  This information is easily queried from the device.

Prior to this commit, I would characterize this device driver as "broken."
It could have been made to work--for varying definitions of "work"--under
the correct circumstances, but it would not have worked well.

At this point, a game controller with a single analog stick and any number
of buttons (up to 32) works pretty much the way it should.  The resulting
button ordering still does not quite make sense to me, but there may be
something else going wrong.  The so-called "axis button" feature is
disabled with this change as I am not sure how to make it work (or if it's
even useful).  Furthermore, it is not yet clear to me how to get back axis
data for more than one stick on a game controller.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@16712 08b38cba-cd3b-11de-854e-f91c5b6e4272