Refactor:	Shift the responsibility for mapping the queried device 		state to button and axis indices from DirectXJoystick to 		DirectXJoystickStandalone.
This change allows gadget::DirectXJoystick to treat the analog axis values
as though they are in a zero-based contiguous block of memory.  Due to the
way that Direct Input works, this is not actually the case, but it is what
we need in order for the analog proxies to behave in a manner that at least
resembles logical.

With this commit, support for the joystick axes and sliders on modern game
controllers is complete and correct.  Further, the button ordering is
sensical and consistent.  The only things missing are the handling of axis
buttons and the "D-Pad" input.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@16716 08b38cba-cd3b-11de-854e-f91c5b6e4272