New feature:	Enable the use of the directional pad(s) (a.k.a, the "hat") 		on a game controller.
With Direct Input, the directional pad value is returned as an angle of
rotation.  The driver reinterprets this into absolute X and Y axis values.
Hence, the directional pad is treated as another joystick.  It can be
reinterpreted further using the axis buttons feature of the driver.

This driver should now be feature-equivalent to the Linux joydev driver.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@16721 08b38cba-cd3b-11de-854e-f91c5b6e4272