Bug fixed:	Adding a new value to a variable-valued property, changing 		its value, and then trying to remove it would result in no 		value removal. SF Bug#:	1106064 (partial)
This changes the manner in which the property sheet's delete button works.
Instead of using the property value to identify which property to remove,
it uses the index of the value in the property's list of values.  This is
safer because the value could change in between the time when the value was
added and when the removal attempt occurs.  However, it would break down
if and when value re-ordering support is added to VRJConfig.  It works fine
for now, though.

NOTE: This fixed the part of SourceForge Bug #1106064 that I reported, but
      it does not fix the initially reported problems.  There is still
      something going wrong with the list refreshing after a value is
      removed.  The correct value is removed, though.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@16775 08b38cba-cd3b-11de-854e-f91c5b6e4272