Bug fixed:	The editor for a property value whose type is a config 		element pointer would sometimes have duplicate entries in 		its list of allowed pointers.
This change makes use of a set instead of a list for the type of mTags.
By definition, a set has at most one of any given entry whereas a list can
have more than one reference to a given entry.  Using a java.util.Set
implementation removes the burden from us to have to keep track of whether
an entry is already in the collection.  However, if a config context
contains multiple elements with the same name, only the first one found
will be displayed in the editor's list of allowed values for the config
element pointer.  As it stands, a config context is not supposed to have
two or more elements with the same name, but this is not yet fully
guaranteed by org.vrjuggler.jccl.config.ConfigContext.

Approved by:	aronb


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@16898 08b38cba-cd3b-11de-854e-f91c5b6e4272