New feature:	Extended the OpenGL frame buffer configuration to allow 		users to provide settings for the auxiliary buffers, the 		stencil buffer, and the accumulation buffer.
This introduces a new version of the config definnition
opengl_frame_buffer_config, but I chose the default property values so
that users of VR Juggler 2.0 Beta 1 or Beta 2 should not be required to
update their configurations unless they want to take advantage of the new
configuration capability.

Bumped version to 1.1.168.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@17002 08b38cba-cd3b-11de-854e-f91c5b6e4272