This is the initial, very much unifinished, version of the DeviceGraph common editor.  This editor is based on JGraph 5.4.3.  It provides a uniform interface for editing the relationships between devices, proxies, and proxy aliases that can be reused by many custom editors.  I have been working on this for quite a while with "manual" revision control, and it is time for some real change tracking to be put in place.
My goal with this editor is to have it replace the bulk of the existing
ProxyEditor code that is based on an out-of-date version of JGraphPad.
Trying to update that editor to the latest JGraphPad version has proven
very difficult due to the massive semantic and syntactic changes made
since JGraphPad 1.0.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@17042 08b38cba-cd3b-11de-854e-f91c5b6e4272