Bug fixed:	Grabbing the desktop border would result in the desktop 		shooting off into infinity. SF Bug#:	832902
This problem arose when gmtl::invert() was optimized to use the inversion
algorithm appropriate for the matrix type.  There is something broken in
this code that prevents gmtl::invert() from working correctly, but I have
never been able to figure out what.  I simply do not have time to track
this down any further, so for now, we will use the slow inversion algorithm
in gmtl::invertFull_orig().  This is how it worked before anyway, so there
is no performance hit.  It would be good if someone would track down the
real problem and back out this revision at some point.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@17082 08b38cba-cd3b-11de-854e-f91c5b6e4272