Refactor:	Make the process of auto-creating device-specific graph 		cells extensible using a functor of sorts.  Now, device 		graph cell creators are registered with GraphHelpers.  My 		long-term hope is that config definitions created by users 		can be handled by instances of 		org.vrjuggler.vrjconfig.commoneditors.DeviceGraph 		auto-magically through the use of Java properties files.
This change required that some private functions in
org.vrjuggler.vrjconfig.commoneditors.devicegraph.GraphHelpers were made
public.  This was not strictly necessary for this specific change, but it is
necessary in order for custom implementations of the interface
org.vrjuggler.vrjconfig.commoneditors.devicegraph.DeviceGraphCellCreator
to be able to make full use of the functionality provided by
org.vrjuggler.vrjconfig.commoneditors.devicegraph.GraphHelpers.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@17131 08b38cba-cd3b-11de-854e-f91c5b6e4272