SF Bug#: [1167555] Keyboard/mouse input in Performer windows broken on X11
Keyboard/Mouse input has been refactored to not use any of Performer's
utility classes. This eliminates the painful multiprocess nightmare.
The downfall is that input events are currently handled in the
APP_CULL_DRAW process. At some point we should consider putting this
in its own thread.

It was a tough game, but I pulled through in the end.
Performer  75
Aron       77 <

Performer input on Win32 needs to be addressed next.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@17242 08b38cba-cd3b-11de-854e-f91c5b6e4272