Tighten up exception handling in doOpen().  When a config file fails to load, call showLoadErrorMessage() (added in the last revision) so that the same dialog box is displayed by doOpen() and doOpenCmdArgs().
This introduces a signficant difference with respect to the previous
behavior when loading a config file that includes other config files.  Now,
the config file(s) that failed to load will not be included in the newly
opened context, but those that loaded successfully will be included.
Previously, one included file failing to load caused the entire context
open operation to fail.  As a result, this change should be tested some
more before it is merged to the 2.0 release branch.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@17761 08b38cba-cd3b-11de-854e-f91c5b6e4272