Fixed an issue with memory being reachable on exit.  Namely, the contents of gadget::ProxyFactory::mConstructors were not being deallocated.  I have changed the collection from using raw pointers to using Boost shared_ptr objects, and I expect that this will probably be controversial. Nevertheless, this is the best option that I could come up with to ensure that gadget::ProxyConstructorBase objects registered with the Proxy Factory--both statically through gadget::ProxyFactory::loadKnownProxies() and dynamically through gadget::ProxyFactory::registerProxy()--would be deleted correctly.
This does introduce a couple of API changes:

   1. Creating an instance of gadget::ProxyConstructor<T> is now done using
      a factory function (gadget::ProxyConstructor<T>::create) rather than
      calling the constructor directly.
   2. gadget::ProxyFactory::registerProxy() now takes a boost::shared_ptr<T>
      instead of a raw pointer.

There may be a side note in the documentation in one or two places about
how the dynamic proxy registration feature exists, but I am sure that it
is not fully explained.  Even if someone were using this feature, s/he
is likely to be more than competent enough to be able to adapt to a change
that would have relatively little impact on existing code.  Furthermore,
the second change above would not affect any existing code since proxy
constructor registration is performed automatically by
gadget::ProxyConstructor<T>.

Bumped version to 1.1.2 to mark this change.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@17796 08b38cba-cd3b-11de-854e-f91c5b6e4272