Extend the vprSingleton* macros so that the singleton object is deleted at application exit and so that singletons implemented using these macros have longevity.  The longevity support comes from Loki's singleton implementation.  In other words, the contents of the files vpr/Util/detail/LifetimeTracker.cpp and vpr/Util/detail/LifetimeTracker.h are from the July 3, 2005, release of the Loki library files Reference/Singleton.cpp and Reference/Singleton.h.  The namespace Loki::Private has been renamed to vpr::detail to fit a little better into VPR.
This change fixes between 40 and 50 cases of heap-allocated memory still
being reachable at the time of VR Juggler application exit--at least on
Linux.  I have been using these patches for about two weeks now and have
tested them on a variety of operating systems and hardware architectures.
I am confident that this change is a positive one, but I am just as
confident that it will expose bugs related to the shutdown of VR Juggler
applications.  In my opinion, finding and fixing those sorts of bugs would
be a very good thing.

The singleton implementation vpr::Singleton<T> has not been extended to
clean up the singleton objects at application exit.  The class could be
modified, but due to the current bugs with that type, it is not seeing
much use anyway.

Bumped version to 1.1.2.

Reviewed by:	No one apparently
Approved by:	Silence on vrjuggler-devel (as usual)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@17866 08b38cba-cd3b-11de-854e-f91c5b6e4272