Refactor:	Removed the call to pfExit() from the method 		vrj::PfDrawManager::closeAPI(). Bug fixed:	Performer-based application objects would never have 		an override of vrj::App::exit() or their destructor called 		as a result of shutting down the kernel.
The basic problem here is two-fold:

   1. An unfortunate side effect of calling pfExit() is that exit(3) is
      called.  Hence, the whole application would exit when
      vrj::PfDrawManager::closeAPI() was called.
   2. The more complete kernel shutdown procedure implemented between the
      2.0 Beta 2 and Beta 3 releases revealed this bug, and users of
      2.0.0 reported that parts of their code stopped being executed.

This behavior of pfExit() basically requires that it be the absolute last
function call in an application.  This means that it has to be at the end
of main() or else anything that should be exected after the invocation of
vrj::PfDrawManager::closeAPI() simply won't be reached.

Prior to 2.0 Beta 3, vrj::DrawManager::closeAPI() was not being called, so
this change is not really so different than the behavior since VR Juggler
was first created.  The main difference is that the recommendation now is
for users to call pfExit() themselves at the end of their main() function.
The documentation has been updated to reflect this.

Bumped version to 2.1.3.

Discussed on:	vrjuggler-info


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@17892 08b38cba-cd3b-11de-854e-f91c5b6e4272