SF Bug#:	1235715 (sort of)
Commented out the call to vpr::Thread::join() in stopSampling(). What is
happening, as far as I can tell, is that the input window is being closed
before stopSampling() is invoked. The call to PostMessage() is failing as a
result. For some reason, though, the call to vpr::Thread::join() is never
returning even though the thread is no longer running (according to the
Visual Studio debugger). Something strange is happening, but at least this
works around the deadlocks on shutdown.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@18210 08b38cba-cd3b-11de-854e-f91c5b6e4272