New feature:	GL/GLX/WGL Extension loading, NV swap_group support
- Added GlExtensionLoader class hierarchy to Draw/GL area of vrj.

This provides a base class that knows how to look up GL/GLX/WGL extensions and the first API specific derived class that is used to load
GLX extensions.  This class provides a way to automatically load, detect, and later call extension functions is a fairly user friendly
way.  In the future, this should be extended to add a windows equivalent.

- Added support for swap_groups.  Initially this uses NV swap_group support for Linux.

This adds a use_swap_group option to the display_system configuration element type.  When this option is true, we attempt to create a swap
group for the local machine and also an associated barrier group so we can sync swaps across a cluster.

Note: This code has not yet been tested with swap lock capable hardware.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@18329 08b38cba-cd3b-11de-854e-f91c5b6e4272